+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated from Java 11 to Java 17
 - Replaced record wrapper by moving its methods to `CsvCallbackHandler`
 - Sealed internal class `AbstractInternalCsvCallbackHandler`
+- Changed implementation of `CsvIndex` and `CsvPage` to Java records
 
 ### Removed
 - Removed deprecated `Limits` class